fix: copy contract, dead code, resource hygiene, and NaN handling
- nan_utils: all-NaN branch now returns copy instead of reference (I2, I9)
- p2108_height_gain: remove unreachable h_m <= 0.0 dead code (I8)
- tile_download_base: replace double-close f.close() with f.flush() (I7)
- macos_compat: accept float timeout values via float() (I10)
- fresnel: use DEFAULT_K_FACTOR constant instead of hardcoded 4.0/3.0 (I11)
- contour/pipeline: add FlushCache() on gdal.Translate before setting to None (I12)
- chart_helpers: replace NaN with empty string in CSV export (I13)
- algorithm/coverage: wrap clutter_grid close() in try/except to preserve original exception (I14)

+++ macos_compat.py
@@ -78,7 +78,7 @@ def _can_spawn(py_exe, env):
         result = subprocess.run(
             [py_exe, "-c", "import encodings"],
             capture_output=True, text=True,
-            timeout=int(os.environ.get("NOWIRES_SPAWN_TIMEOUT", "5")),
+            timeout=float(os.environ.get("NOWIRES_SPAWN_TIMEOUT", "5")),
             env=env,
         )
         return result.returncode == 0
